convert resource_file to resource shortcodes
This refactors `markdown_cleanup_baseurl` management command to be `markdown_cleanup` which can now be passed a rule.
