+++ websites/management/commands/markdown_cleaning/baseurl_rule.py
@@ -2,16 +2,10 @@
 import importlib
 import os
 import re
-from contextlib import ExitStack
 from typing import Iterable
 
-from django.core.management import BaseCommand
-from django.core.management.base import CommandParser
-from django.db import transaction
-
-from websites.management.commands.util import (
-    WebsiteContentMarkdownCleaner,
-    progress_bar,
+from websites.management.commands.markdown_cleaning.cleanup_rule import (
+    MarkdownCleanupRule,
 )
 from websites.models import WebsiteContent
 
@@ -67,8 +61,14 @@ class ContentLookup:
             return self.website_contents[(website_id, dirpath, filename)]
 
 
-class BaseurlReplacer:
-    """Replacer function for use with WebsiteContentMarkdownCleaner. Replaces
+def get_all_website_content():
+    return WebsiteContent.all_objects.all().only(
+        "dirpath", "filename", "markdown", "website_id"
+    )
+
+
+class BaseurlReplacementRule(MarkdownCleanupRule):
+    """Replacement rule for use with WebsiteContentMarkdownCleaner. Replaces
     baseurl links with < resource_link > shortcodes.
 
     This is intentially limited in scope for now. Some baseurl links, such as
@@ -76,10 +76,13 @@ class BaseurlReplacer:
     replacement.
     """
 
-    baseurl_regex = r"\[(?P<title>[^\[\]\n]*?)\]\({{< baseurl >}}(?P<url>.*?)\)"
+    regex = r"\[(?P<title>[^\[\]\n]*?)\]\({{< baseurl >}}(?P<url>.*?)\)"
 
-    def __init__(self, content_lookup: ContentLookup):
-        self.content_lookup = content_lookup
+    alias = "baseurl"
+
+    def __init__(self):
+        website_contents = get_all_website_content()
+        self.content_lookup = ContentLookup(website_contents)
 
     def __call__(self, match: re.Match, website_content: WebsiteContent):
         original_text = match[0]
@@ -95,63 +98,3 @@ class BaseurlReplacer:
             )
         except KeyError:
             return original_text
-
-
-class Command(BaseCommand):
-    """
-    Replaces baseurl-based links in markdown with < resource_link > shortcodes.
-    """
-
-    help = __doc__
-
-    def add_arguments(self, parser: CommandParser) -> None:
-        parser.add_argument(
-            "-o",
-            "--out",
-            dest="out",
-            default=None,
-            help="If provided, a CSV file of baseurl-based links and their replacements will be written to this path.",
-        )
-        parser.add_argument(
-            "-c",
-            "--commit",
-            dest="commit",
-            default=False,
-            help="Whether the changes to markdown should be commited. The default, False, is useful for QA and testing when combined with --out parameter.",
-        )
-        super().add_arguments(parser)
-
-    def validate_options(self, options):
-        """Validate options passed to command."""
-        if not options["commit"] and not options["out"]:
-            raise ValueError("If --commit is falsy, --out should be provided")
-
-    def handle(self, *args, **options):
-        self.validate_options(options)
-        commit = options["commit"]
-        out = options["out"]
-
-        with ExitStack() as stack:
-            wc_list = WebsiteContent.all_objects.all().only(
-                "dirpath", "filename", "markdown", "website_id"
-            )
-            if commit:
-                stack.enter_context(transaction.atomic())
-                wc_list.select_for_update()
-
-            content_lookup = ContentLookup(wc_list)
-            replacer = BaseurlReplacer(content_lookup)
-            cleaner = WebsiteContentMarkdownCleaner(
-                BaseurlReplacer.baseurl_regex, replacer
-            )
-
-            wc: WebsiteContent
-            for wc in progress_bar(wc_list):
-                cleaner.update_website_content_markdown(wc)
-
-            if commit:
-                wc_list.bulk_update(cleaner.updated_website_contents, ["markdown"])
-
-        if out is not None:
-            outpath = os.path.normpath(os.path.join(os.getcwd(), out))
-            cleaner.write_matches_to_csv(outpath)
